update triples code
Former-commit-id: 1ebd023c3c1532e3543aa3668ac3bb1f4fcc3b28
Former-commit-id: 457618be8d16e73fb231ffbee8980183e7f49fb9
Former-commit-id: b539f257139cf9942a2962cbcd51fa43abcd12a3
Former-commit-id: 9b6365e425af5f502d71f3ef4e436a900c7e9126 [formerly 5dfd8be85b35311ecfa890e9b369be1306c7abc2]
Former-commit-id: b980197c320005e95f0cff7154626f7792d1941e

+++ methods/cc/ccsd_t/ccsd_t_all_fused_cpu.cpp
@@ -16,13 +16,13 @@ void total_fused_ccsd_t_cpu(size_t base_size_h1b, size_t base_size_h2b, size_t b
                             size_t size_d2_t2_all, size_t size_d2_v2_all,
                             size_t size_s1_t2_all, size_t size_s1_v2_all,
                             // 
-                            size_t* list_d1_sizes, 
-                            size_t* list_d2_sizes, 
-                            size_t* list_s1_sizes, 
+                            int* list_d1_sizes, 
+                            int* list_d2_sizes, 
+                            int* list_s1_sizes, 
                             // 
-                            std::vector<size_t> vec_d1_flags,
-                            std::vector<size_t> vec_d2_flags,
-                            std::vector<size_t> vec_s1_flags, 
+                            std::vector<int> vec_d1_flags,
+                            std::vector<int> vec_d2_flags,
+                            std::vector<int> vec_s1_flags, 
                             // 
                             size_t size_noab, size_t size_max_dim_d1_t2, size_t size_max_dim_d1_v2,
                             size_t size_nvab, size_t size_max_dim_d2_t2, size_t size_max_dim_d2_v2,
